ai_mock_interviewer/requirements.txt
# Core dependencies
pydantic>=2.9
python-dotenv>=1.0

# AI/ML
google-genai>=1.59.0

# HTTP client
aiohttp>=3.11
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=8.3
pytest-asyncio>=0.24
